Handle variable names and indexes within quoted contexts
There are a lot of changes in this commit but the main one is support for
tokenizing variable names + an optional index within a double-quoted context.

This is really tricky it necessitates parsing a `VariableName` token type
instead of just returning it as text and leaving it for the parser to handle,
as otherwise we are not going to catch the index that comes after the variable
name in a quoted context.