Rework fd detection & split tokenizer tests to multiple files
We no longer limit an fd reference to come immediately after a redirection
symbol (e.g. `echo foo>&bar`) in order to mimic fish's behavior. While fish does
not accept `&2` as an fd in `echo foo> &2` since it's not glued to the
redirection operator, it doesn't interpret it as text either. We now do the
same.

The incremental recompilation was taking a hit with the larger file size, and
tbh, it was a little too long to keep track of. Putting the tests in their own
modules means that there's some duplication now between the module name and the
test name, but we can resolve that later or leave it as-is (most likely the
latter).