Update ampersand logic and add more tests
It's not sufficient to determine that we don't have a fragment, as we could have
broken in the middle of a text token (thanks to an escape). Record the type each
time a token is returned and check it to make sure we're not in the middle of a
text token.