Add `UnifiedTokenIterator` and corresponding trait
Allows using `tokenize(b"...").unified()` to return a less-optimized but
easier-to-use interface that combines adjacent `TokenType::Text` tokens (not
separated by whitespace or operators).

This adds an allocation in cases where there is more than one such token text
token in a row so that their text can be returned in one contiguous chunk.