Don't tokenize start/end quotes
This removes the token types `TT:SingleQuote` and `TT:DoubleQuote`; the contents
of a quoted string are returned as text, but the parser doesn't care about the
presence of the quotes themselves - they are purely tokenizer constructs.

If we need to convey to the shell how the text was quoted it would be better to
attach to `TokenType::Text` some state "TextType" or whatever that could be
"literal", "single-quoted", "double-quoted", or "escaped", as the case may be.

This simplifies the parsing loop by minimizing the number of states that need to
be handled - now all text can be handled the same, regardless of how it was
formed.