Change how we handle partial strings
The approach we were using to avoid allocating on escapes and quotes was all
nice in theory but it wasn't very workable. The problem arose from non-text
tokens that could contain quoted or escaped characters (such as variable names).
These would cause the parser to see either a `TT::VariableName` followed by
`TT:Whitespace` or at best `TT::VariableName` twice, which doesn't make sense.
*Someone* is going to have to handle the concatenation at some point for some
reason (e.g. to convert to args for a process, to convert into an internal
variable name key, etc) so it might as well be the tokenizer.